remove extra fields from AbstractTreeNode to avoid potential memory estimation problems with ReadTreeNode
